+++ NEWS.md
@@ -10,6 +10,7 @@ CRSSIO 0.8.2.9000
   - also updated internal historical index data to go through 2019. 
 - added `as_nfd.data.frame()`, `as_nfd.list()`, `as_crss_nf.data.frame()`, and `as_crss_nf.list()` to convert data frames to `nfd` and `crss_nf` objects. (#82)
 - added Meko et al. (2017) and Woodhouse et al. (2006) Lees Ferry paleo reconstrutions for use in package. (#114)
+- Fixed bug in `plot.nfd()` so that it correctly errors, and posts a helpful error message, if `"cloud"` or `"spaghetti"` plots are specified for monthly `nfd` data. (#110)
 
 CRSSIO 0.8.2
 =================